4-Knowledge Base/requirements.txt
farm-haystack==0.9.0
farm==0.8.0